modularize classifier and generator
